requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
datasets==3.5.0
decorator==4.4.2
dill==0.3.8
docopt==0.6.2
exceptiongroup==1.2.2
ffmpeg-python==0.2.0
filelock==3.17.0
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.55.8
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
google-api-core==2.24.1
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
Js2Py==0.74
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.0
more-itertools==10.6.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.4.2
noisereduce==3.0.3
numba==0.61.0
numpy==2.1.3
oauth2client==4.1.3
oauthlib==3.2.2
openai-whisper==20240930
package_name==0.1
packaging==24.2
pandas==2.2.3
pillow==10.4.0
pipwin==0.5.2
platformdirs==4.3.6
pooch==1.8.2
proglog==0.1.10
propcache==0.3.1
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyDrive==1.3.1
pydub==0.25.1
pyjsparser==2.7.1
pyparsing==3.2.1
PyPrind==2.11.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.16.1
pySmartDL==1.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytube==15.0.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
replicate==1.0.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
SpeechRecognition==3.14.2
srt==3.5.3
sympy==1.13.1
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
tqdm==4.67.1
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.3.0
vosk==0.3.45
waitress==3.0.2
websockets==14.2
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.0
youtube-dl==2021.12.17
yt-dlp==2025.3.25
